Extract _invoke_command into api_helpers.py
